pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "operation-agent"
version = "0.1.0"
description = "障害診断エージェント - Strands Agents SDK + Amazon Bedrock"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.30.0",
    "strands-agents-tools>=0.2.22",
    "boto3>=1.38.0",
    "pydantic-settings>=2.0",
    "botocore[crt]>=1.42.68",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "pytest>=8.0",
    "jupyterlab>=4.0",
    "ipykernel>=6.0",
    "lefthook>=2.1.4",
    "mypy>=1.10",
    "boto3-stubs[lambda]>=1.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: integration tests requiring AWS credentials",
]
